src/python/requirements.txt
fastapi>=0.101.1
uvicorn
weaviate-client>=4.5.0
transformers>=4.36.2
neo4j>=5.11.0
torch
python-dotenv>=1.0.0
adapters>=0.1.2
starlette>=0.27.0

numpy
pandas
unidecode
nltk
blingfire
awscli
lightgbm

frozendict
